+++ packages/piveau-hub-ui-modules/lib/datasetDetails/distributions/distributionDetails/DistributionContent.vue
@@ -62,10 +62,10 @@
           {{ filterDateFormatEU(distribution.modificationDate) }}
         </td>
         <td v-else-if="has(distribution, 'releaseDate') && !isNil(distribution.releaseDate)">
-          {{ filterDateFormatEU(distribution.modificationDate) }}
+          {{ filterDateFormatEU(distribution.releaseDate) }}
         </td>
         <td v-else>
-          --
+          Unbekannt
         </td>
       </tr>